+++ lib/cluster_murmur/application.ex
@@ -3,13 +3,20 @@ defmodule ClusterMurmur.Application do
 
   use Application
 
+  alias ClusterMurmur.Runtime.{ProductionApplication, RootSupervisor}
+
   @impl true
   def start(_type, _args) do
-    children = [ClusterMurmur.Repo]
+    with {:ok, children} <- child_specs() do
+      RootSupervisor.start_link(children)
+    end
+  end
 
-    Supervisor.start_link(children,
-      strategy: :one_for_one,
-      name: ClusterMurmur.Supervisor
-    )
+  defp child_specs do
+    case Application.fetch_env(:cluster_murmur, :standalone_runtime) do
+      {:ok, true} -> ProductionApplication.child_specs()
+      {:ok, false} -> {:ok, [ClusterMurmur.Repo]}
+      _invalid -> {:error, :invalid_application}
+    end
   end
 end